Add a super basic script for packaging release zips
This will allow us to easily create a zip of the entire project,
including the vendor folder. Going forward we can make this more
advanced so that we can supply it a version number etc and it will
build a proper ‘release’ build, but for now it at least takes away some
tedium involved in packaging the project for people on shared hosting.